Run Chooser integration tests for the new features
Add a new test parameter, feature flag set, and run Chooser tests for
two set values: all flags are off and all flags are on. This ensures
that our integration tests cover the new features.
A new test rule is added to ignore tests that were designed for a
specific flag values when running in a set that does not have those
values set.

Bug: 258838272
Test: the modified test itself
Change-Id: If646ef123a383e801fda55d601e10b186c6c5c1f