Remove plain `startActivity` plumbing APIs
These were unused, and the pruned-down API surface better indicates
that we *always* launch activities to an explicit *user* (and *maybe*
with additional specification about the caller identity -- something
to do with the "safe forwarding mode" feature I'm still trying to
get my mind aroud).

While I was in here, I changed the signature of the one 'activity
starter' method that arbitrarily took a `ResolverActivity` instead of
the more generic `Activity` used in the others. We've been commenting
about the opportunity to make this fix for a while now, and there's
no time like the present.

Test: `atest IntentResolverUnitTests`
Bug: 202167050
Change-Id: I33be96b68e6feba8b7fdb0295431373aadd2c2bf