Remove UserHandle ivar from ResolverListController.
ResolverListController instances are 1:1 with the ResolverListAdapter
instances that hold them, and the adapters already carry their
associated UserHandles, so this was just one more "source of truth"
to potentially get out of sync as we're trying to clean up management
of these handles throughout the project. (And by exposing the handle
in a now-removed getter, it also added a redundant mechanism for
clients to access this value, making it harder to track down all the
usages in our application.)

I also took this opportunity to factor out a lot of boilerplate in
our tests to use a helper. That's not generally considered a good
practice for testing, but in this case the helper already existed
and just wasn't used in some cases -- where we instead duplicated
verbose boilerplate that represents less about the conditions we'd
actually expect to see in practice. Long-term we should clean up a
*lot* of details in our test design, so I think for now it's OK to
just use a more terse version of the design we already have.

Eventually I'd like to break up ResolverListController into a few
components with narrower focus, so removing any unnecessary
responsibilities is a great start.

Test: `atest IntentResolverUnitTests`
Bug: 202167050
Change-Id: Iade5b8d1f4e31d5439c234fbb1d82169ec01a386