Always terminate Chooser after a refinement flow.
In normal (non-refinement) target selection, if the user selects a
suspended target, we attempt to launch it anyways in order to prompt
a system message dialog, but we suppress our normal post-launch
termination so that the user can try to select a different target.

We had similar "bounce-back" logic if a post-refinement launch failed
as a result of a suspended target, but the UX doesn't really make
sense in the refinement case (and I'm also not 100% confident that our
implementation would be technically equipped to handle the re-entry).

I've removed the "bounce-back" suppression in the post-refinement
case, so if we show the dialog about the package being suspended,
that's the end of the session. (Although note, unless the package was
suspended *during* refinement, we would've already shown the dialog
instead of going to refinement in the first place.)

Bug: 273864843
Test: manually tested behavior around suspended targets, and ran
CTS to verify no regressions in normal refinement usage.

Change-Id: I4e57add285224b7f311c4d0532168dc984ee331b