Terminate refinement on back-out w/ no result
Bug: 273864843
Test: manually tested back-out behavior with a custom refinement
activity that calls `finish()` on launch. Ran CTS &
`IntentResolverUnitTests` to confirm no regressions in the normal
refinement flow. (Also noted that this condition *isn't* triggered
when we finish after sending a valid RESULT_OK refinement. That's not
really required for correctness, but it's still simpler/preferable
that way, so it was a relief to confirm.)

Change-Id: I165c958e3633430006b7977faa7617a3f87b1092