Show preview content in landscape.
Make stuff smaller and less tall on landscape non-tablets.
Let the system restart sharesheet on rotation to allow for resources to
change.

Tablet displays in landscape generally should follow the un-compressed
UI of portrait phones, so the resources for that configuration have to
be listed twice.

This doesn't dismantle any of the code for handling orientation changes
ourselves, as taking that apart may be error-prone.

This is meant to a first pass, additional landscape miniturization will
follow, as there still isn't a ton of room to scroll at the bottom.

Bug: 267522604
Test: Testing in portrait and landscape with a variety of ShareTest
      payloads
Change-Id: I4697ad9510526cc9ce5f9e0630080bef5fad377c