Move Nearby to the front of the app list.
Introduce a concept of a "cemented" component that always goes to the
front of the list. Just allowing up to one of these since we really
don't expect to ever have more than one.

Remove the nearby action code as well, along with some associated
cleanup.

Not in this CL: Prevent this target from being pinned.

Bug: 275787600
Test: atest com.android.intentresolver
Change-Id: I55c551f930de2d642cd0e5a42c6cdbd6e1918f0b