Allow refinement of more non-filter-equals fields.
This change allows callers' "refinement activities" to override the
source intent's ClipData and given 'extras' entries. Note neither of
these fields are considered as factors in `Intent.filterEquals()`, so
allowing callers to refine their values doesn't violate the documented
guardrails on refinement.

It also makes sense to support, as otherwise refinement wouldn't be
able to modify the "payload" share content at all, which (by the docs
for refinement) is clearly something we'd *intended* to support.

Test: CTS & atest. Upcoming CL adds new CTS coverage for payload
refinements (already verified against this CL).
Bug: 262808892

Change-Id: I7fb344fadc04f5ebed83f70f3b16ba9c1f28391c