Remove auto-launching for single target, multiple tabs
This can lead to poor UX when auto-launching across tabs.
If the intent picker should be shown (e.g dev calls startActivity),
and there is only a single app to handle the intent, then this has no
affect. It is handled in either IntentForwarderActivity, or in
ResolverActivity (mini-resolver).

However, if the dev calls Intent.createChooser/ACTION_CHOOSER, without
an ACTION_SEND intent (ChooserActivity), then we show something
that looks like an intent picker, which has this incorrect auto-launching logic.

For now we simply remove the logic because it doesn't make sesnse
in an actual sharing context, and if ACTION_SEND is not used it leads
to a security vulnerability and very poor UX (b/270700718). In a future
release we can hopefully improve this and treat this case in-line with
the intent-picker (e.g., showing the mini-resolver).

Bug: 272208024
Bug: 270700718
Test: atest ResolverActivityTest
Test: atest UnbundledChoserActivityTest
Test: Manually tested, attached screenshots in bug
Change-Id: I1a340cab94230cada1e7e805195070b032e7c5ff