Make app-provided chooser targets only show in the default profile
Previously if, say, a work-profile app injected a chooser target, that
target would show in *both* profiles, but showing it in the opposite
(personal) profile doesn't really make sense since and uses up direct
share space.

The placement of this code is highly-suspect but I'm avoiding changing
that in this CL given the proximity to release.

Bug: 280517912
Test: atest UnbundledChooserActivity
Change-Id: Ia73699cde6347cd300eae388a42a2433b3840d63