Allow chooser targets to use two lines
Previous behavior is that direct share would get two lines
but app targets got one line for the label and one line for
the sub-label.

This forced truncation of labels even when no sub-label was
present (see bug).

This change evolves that by making the following changes:
1. Set the maxLines of the label instead of lines directly.
2. If there's no sublabel, set maxLines to 2.
3. Change the grid item's gravity so they're top-aligned
   (previously they were all the same height yet centered).

Setting the line count directly didn't work because the
compound drawable would always be drawn as if there were two
lines.

One subtlety here is that a row of targets that all fit
within a single line will have less space between icons.

Bug: 280289219
Test: atest UnbundledChooserActivityTest
Test: manual review how a variety of targets render.
Change-Id: Id9831a310738631c9f767827fec595117ac3a487