Creates ChooserActivity alias
This alias allows us to avoid a bug that caused Share Sheet to crash any
app that tried to use it. It also removes the previous solution as it
had the potential to not take effect if too many receivers tried to act
on the BOOT_COMPLETE broadcast.

Test: $ adb root
Test: $ adb shell pm disable com.android.intentresolver/.ChooserActivity
Test: $ adb shell pm resolve-activity -a android.intent.action.CHOOSER
Test: Verify that the activity resolves
BUG: 283722356
FIX: 283722356
Change-Id: I820c5f88204c83e635cb877b222bdf6cbe53aaff
Merged-In: I820c5f88204c83e635cb877b222bdf6cbe53aaff