Add headline view argument to content preview UI
"Sticky" headline is one of the requirements of the scrollable preview
feature. Currently headling row is included in every preview type we
have and to be "sticky" it should be moved out of all of them. This
change is a preparation work that makes all of the preview UI
controllers aware of a possible external headline row (but no actual
external headline view is used).

Bug: 287102904
Test: manual testing of all preivew types checking that there are no
regressions
Test: atest com.android.intentresolver.contentpreview

Change-Id: Ib6110aaa82246e3354fdce18f431ab1c116e7b73