Merge the `MultiProfilePagerAdapter` base classes
That is, this CL consolidates `GenericMultiProfilePagerAdapter` up
into the base `AbstractMultiProfilePagerAdapter` (and renames to drop
the prefixes). This advances an in-progress refactoring to simplify
these classes (as in an earlier `GenericMultiProfilePagerAdapter`
TODO comment when that class was first introduced).

This prepares for other upcoming work in the base class (especially in
service of "private space" support) since the responsibilities were
unclear and arbitrarily split between these two bases, which had no
other direct clients/subclasses (i.e., this CL is exactly equivalent
to the earlier code, but removes a class from the hierarchy).

I've also started a little bit of cleanup around method visibility but
left much of that work out-of-scope for now.

Test: IntentResolverUnitTests / CtsSharesheetDeviceTest
Change-Id: I073e0bec5764ea16736af585af0bc9f744089d03