Defer finish() to onStop() when doing shared transition
We used to immediately finish() after starting the shared element
transition. In high-latency situations, this transition took a while to
start and resulted in some broken UI in window manager. Instead of
finishing immediately, we finish onStop() such that we can wait for the
transition to actually begin.

Bug: 292158050
Test: Share an image that has high latency to load. Validate that the UI
looks normal (albeit slow) when the edit transition happens (see bug).

Change-Id: I66686398781aeea630c500c48198c5febc1e8fc2