Begin extracting `emptystate` module.
This CL just contains the initial code-moves to group relevant classes
(including some inner classes/interfaces that are being pulled out of
`MultiProfilePagerAdapter`), without any other anticipated fixes or
reorganization. This corresponds to snapshots 1-3 of the prototype
ag/24516421 (to be followed with the other work that was started in
that CL).

Bug: 302311217
Test: IntentResolverUnitTests, CtsSharesheetDeviceTest
Change-Id: If82a761193f6ff0605c5a46c106f7c95699350c0