Starts splitting the Chooser and Resolver logic
Creates two new interfaces that house previous activity logic:
CommonActivityLogic and ActivityLogic, which extends it. Also creates
a concrete instantiation of CommonActivityLogic and two concrete
instantiations of ActivityLogic (one for each activity). The concreate
ActivityLogic classes delegate to the concrete CommonActivityLogic for
that interface surface.

The intent is for these classes to act as a scaffolding for separating
the Chooser and Resolver activities.

Test: atest com.android.intentresolver.v2
BUG:302113519
Change-Id: I82084aad2d7f759763cf6a77c1bebbb3d3f82c5c